notifications/scheduler/requirements.txt
psycopg==3.1.18
psycopg-binary==3.1.18
pika==1.3.2
schedule==1.2.2
pydantic==2.8.2
pydantic-settings==2.4.0
psycopg2-binary==2.9.9
